+++ src/MuMu_RichPresence/MuMu/Interop/AnonymousAsyncDisposable.cs
@@ -1,0 +1,10 @@
+﻿namespace Dawn.MuMu.RichPresence.MuMu.Interop;
+
+public class AnonymousAsyncDisposable(Func<ValueTask> dispose) : IAsyncDisposable
+{
+    public async ValueTask DisposeAsync()
+    {
+        GC.SuppressFinalize(this);
+        await dispose();
+    }
+}