+++ src/MuMu_RichPresence/MuMu/Interop/AppInfo.cs
@@ -1,0 +1,6 @@
+﻿namespace Dawn.MuMu.RichPresence.MuMu.Interop;
+
+public readonly record struct AppInfo(
+    string PackageName,
+    int Pid,
+    DateTimeOffset StartTime);