+++ src/MuMu_RichPresence/MuMu/Interop/ConnectionInfo.cs
@@ -1,0 +1,10 @@
+﻿namespace Dawn.MuMu.RichPresence.MuMu.Interop;
+
+public readonly record struct ConnectionInfo(
+    string LocalIP,
+    int LocalPort,
+    string ADBPath
+)
+{
+    public const int FALLBACK_PORT = 5555;
+}