+++ src/MuMu_RichPresence/MuMu/Interop/IMuMuInterop.cs
@@ -1,0 +1,8 @@
+﻿namespace Dawn.MuMu.RichPresence.MuMu.Interop;
+
+public interface IMuMuInterop
+{
+    public Task<AppInfo> GetForegroundAppInfo();
+
+    public Task<string> GetInfo(AppInfo info, string path);
+}